feat: provider-level options, useWhenProviderReady
* adds the ability to configure options at the context-provider level
* adds new useWhenProviderReady hook which force-suspend components until the provider is ready (especially good for React 17 shortcomings
