+++ packages/react/src/provider/index.ts
@@ -1,1 +1,3 @@
 export * from './provider';
+export * from './use-open-feature-client';
+export * from './use-when-provider-ready';
